+++ console/src/reducers/index.js
@@ -11,6 +11,7 @@ import {reducers as wizard} from "../process/wizard";
 import {reducers as portal} from "../process/portal";
 import {reducers as project} from "../project";
 import {reducers as projectList} from "../project/list";
+import {reducers as repository} from "../project/repository";
 import {reducers as secret} from "../user/secret";
 import {reducers as modal} from "../shared/Modal";
 
@@ -25,6 +26,7 @@ export default {
     portal,
     project,
     projectList,
+    repository,
     secret,
     modal,