+++ server/impl/src/main/java/com/walmartlabs/concord/server/console/ConsoleService.java
@@ -1,8 +1,11 @@
 package com.walmartlabs.concord.server.console;
 
 import com.walmartlabs.concord.server.project.ProjectDao;
+import com.walmartlabs.concord.server.project.RepositoryManager;
 import com.walmartlabs.concord.server.security.UserPrincipal;
 import com.walmartlabs.concord.server.security.ldap.LdapInfo;
+import com.walmartlabs.concord.server.security.secret.Secret;
+import com.walmartlabs.concord.server.security.secret.SecretManager;
 import org.apache.shiro.SecurityUtils;
 import org.apache.shiro.authz.annotation.RequiresAuthentication;
 import org.apache.shiro.subject.Subject;
@@ -11,18 +14,24 @@ import org.sonatype.siesta.Resource;
 import javax.inject.Inject;
 import javax.inject.Named;
 import javax.ws.rs.*;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.Response.Status;
 
 @Named
 @Path("/api/service/console")
 public class ConsoleService implements Resource {
 
     private final ProjectDao projectDao;
+    private final RepositoryManager repositoryManager;
+    private final SecretManager secretManager;
 
     @Inject
-    public ConsoleService(ProjectDao projectDao) {
+    public ConsoleService(ProjectDao projectDao, RepositoryManager repositoryManager, SecretManager secretManager) {
         this.projectDao = projectDao;
+        this.repositoryManager = repositoryManager;
+        this.secretManager = secretManager;
     }
 
     @GET
@@ -33,13 +42,13 @@ public class ConsoleService implements Resource {
         Subject subject = SecurityUtils.getSubject();
         if (subject == null) {
             throw new WebApplicationException("Can't determine current user: subject not found",
-                    Response.Status.INTERNAL_SERVER_ERROR);
+                    Status.INTERNAL_SERVER_ERROR);
         }
 
         UserPrincipal u = (UserPrincipal) subject.getPrincipal();
         if (u == null) {
             throw new WebApplicationException("Can't determine current user: entry not found",
-                    Response.Status.INTERNAL_SERVER_ERROR);
+                    Status.INTERNAL_SERVER_ERROR);
         }
 
         LdapInfo i = u.getLdapInfo();
@@ -69,5 +78,37 @@ public class ConsoleService implements Resource {
     @RequiresAuthentication
     public boolean isProjectExist(@PathParam("projectName") String projectName) {
         return projectDao.exists(projectName);
+    }
+
+    @POST
+    @Path("/repository/test")
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Produces(MediaType.APPLICATION_JSON)
+    @RequiresAuthentication
+    public boolean testRepository(RepositoryTestRequest req) {
+        try {
+            Secret secret = secretManager.getSecret(req.getSecret());
+            repositoryManager.testConnection(req.getUrl(), req.getBranch(), req.getCommitId(), secret);
+            return true;
+        } catch (Exception e) {
+            String msg;
+            Throwable t = e;
+            while (true) {
+                msg = t.getMessage();
+                t = t.getCause();
+                if (t == null) {
+                    break;
+                }
+            }
+
+            if (msg == null) {
+                msg = "Repository test error";
+            }
+
+            throw new WebApplicationException(Response.status(Status.INTERNAL_SERVER_ERROR)
+                    .header(HttpHeaders.CONTENT_TYPE, MediaType.TEXT_PLAIN)
+                    .entity(msg)
+                    .build());
+        }
     }
 }