+++ server/impl/src/main/java/com/walmartlabs/concord/server/project/RepositoryException.java
@@ -2,6 +2,10 @@ package com.walmartlabs.concord.server.project;
 
 public class RepositoryException extends Exception {
 
+    public RepositoryException(Throwable cause) {
+        super(cause);
+    }
+
     public RepositoryException(String message) {
         super(message);
     }